refactor: move admin components into admin folder
