+++ site/src/pages/AdminPage/Reports/ReportGroup.scss
