+++ site/src/pages/AdminPage/Verify/UnverifiedReview.scss
@@ -1,0 +1,6 @@
+.verification-buttons {
+  display: flex;
+  flex-direction: row;
+  gap: 12px;
+  margin: 0px 8px;
+}
