+++ site/src/pages/AdminPage/Verify/Verify.scss
@@ -5,10 +5,3 @@
   }
   border-radius: var(--border-radius);
 }
-
-.verification-buttons {
-  display: flex;
-  flex-direction: row;
-  gap: 12px;
-  margin: 0px 8px;
-}
